requirements.txt
Celery
Django==4.1.6
django-extensions==3.2.1
ipython
Jinja2
django-debug-toolbar
django-crispy-forms
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
crispy-bootstrap5
requests


#flake8
flake8
flake8-builtins
flake8-print

pytest
pytest-django
pytest-mock
